Correcting Myself in the Readme and Adding Extra Effects
In hindsight, I realized that I made a tiny mistake in my initial read
file when i stated that the pageYOffset property was that o an element
(which I realized only a day after when I re-read the README doc. I
also tied the left & right properties of the elements so they now not
only move vertically buy horizontally in different speeds.